requirements.txt
google-generativeai
flask
python-dotenv
retrying
google-api-python-client
google-auth
google-auth-oauthlib
pillow